Create a database key that will sort by time for the block files
By prefixing the block hash with the blocktime plus a ":" separator
the records will now be sorted by time on disk. This has a large
perforance benefit in reducing the frequency and scope of database
compactions.  The only files that now undergo compaction are the most
recent one created, rather than any and all files as was the previous
case.