Fix the downgrade process from blockdb to sequential files
Was not using pindex->pprev when looking up undo files.

Also the nStatus was not getting removed when a block or undo file
didn't exist or couldn't get read from the blockdb. This was causing
us to set every header as though it had block data and undo files
when in fact it didn't yet. As a result we we're not able to continue
with a sync if it hadn't yet finished.