Extended Version Message implementation
This is a proposal and implementation for augmenting the `version`
message in the BCH protocol with an `xversion` message to provide
a key value map of 64bit integers to `std::vector<uint8_t>` values
(called `xmap`) that can be used to signal general configuration
and version options to connected peers.

This implementation comes with a specification document in
doc/xversionmessage.md, a simple unit test and a fuzzer test
case. It also comes with a proposal on how to organize and
name the `xmap` key space between implementations, as well as a tool
to process a key space definition file (xversionkeys.dat) into
a C++ header.

Furthermore, the BU-specific BUVERSION message has been retired (but
is still handled) and replaced with an appropriate entry in this map.

Thanks to everyone who gave me input on this PR:

@gandrewstone, @Greg-Griffith, @dgenr8, @sickpig, @dagurval,
@jasonbcox, @schancel, @deadalnix